Resource de entidad intermedia con pk

+++ src/main/java/sv/edu/ues/occ/ingenieria/tpi_2026/ingreso/dto/AspiranteIdentificacionDTO.java
@@ -13,7 +13,7 @@ import jakarta.validation.constraints.NotNull;
  */
 public record AspiranteIdentificacionDTO(
         @NotNull
-        int idAspirante,
+        Long idAspirante,
         @NotNull
         int idTipoIdentificacion,
         @NotBlank
